+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.11.0 -- 2023-??-??
 
 * Add `Data.Bytes.Parser.Latin.hexWord32`.
+* Add `Data.Bytes.Parser.Latin.hexFixedWord(128|256)`.
 
 ## 0.3.10.0 -- 2023-07-25